fixing / reorganizing react router dom
