+++ frontend/package.json
@@ -15,6 +15,7 @@
     "prop-types": "^15.8.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-icons": "^4.10.1",
     "react-router-dom": "^6.12.1",
     "vite": "^4.0.4"
   },
