--- frontend/src/App.jsx
@@ -1,35 +1,0 @@
-/* eslint-disable import/no-extraneous-dependencies */
-import { createBrowserRouter, RouterProvider } from "react-router-dom";
-import Home from "./pages/Home";
-import ErrorPage from "./pages/ErrorPage";
-import LoginPage from "./pages/LoginPage";
-import "./App.css";
-import DecisionItem from "./pages/DecisionItem";
-
-const router = createBrowserRouter([
-  {
-    path: "/",
-    element: <Home />,
-    errorElement: <ErrorPage />,
-  },
-  {
-    path: "/login",
-    element: <LoginPage />,
-    errorElement: <ErrorPage />,
-  },
-  {
-    path: "/decision",
-    element: <DecisionItem />,
-    errorElement: <ErrorPage />,
-  },
-]);
-
-function App() {
-  return (
-    <div className="App">
-      <RouterProvider router={router} />
-    </div>
-  );
-}
-
-export default App;
