+++ frontend/src/main.jsx
@@ -1,12 +1,40 @@
+/* eslint-disable import/no-extraneous-dependencies */
 import React from "react";
+import { createBrowserRouter, RouterProvider } from "react-router-dom";
 import ReactDOM from "react-dom/client";
+import ErrorPage from "./pages/ErrorPage";
+import "./App.css";
+import DecisionItem from "./pages/DecisionItem";
+import LoginPage from "./pages/LoginPage";
+import Root from "./routes/Root";
+import NavBar from "./components/NavBar";
 
-import App from "./App";
+const router = createBrowserRouter([
+  {
+    path: "/",
+    element: <Root />,
+    children: [
+      {
+        path: "/decision",
+        element: <DecisionItem />,
+        errorElement: <ErrorPage />,
+      },
+      {
+        path: "/login",
+        element: <LoginPage />,
+        errorElement: <ErrorPage />,
+      },
+      {
+        path: "/nav",
+        element: <NavBar />,
+        errorElement: <ErrorPage />,
+      },
+    ],
+  },
+]);
 
-const root = ReactDOM.createRoot(document.getElementById("root"));
-
-root.render(
+ReactDOM.createRoot(document.getElementById("root")).render(
   <React.StrictMode>
-    <App />
+    <RouterProvider router={router} />
   </React.StrictMode>
 );
